Silence warnings
- warnings 'Set but not used [-Wunused-but-set-parameter]'
- warnings for variable uninitialized
- warnings: implicit conversion from 'long' to 'double'
- warnings: first argument in call to 'memset' is a pointer to non-trivially copyable type 'Symbol'
- warning: ignoring duplicate libraries
- warning: Unused parameter

Silence warning with (void)variable

[IA Generated]

In C/C++, writing (void)variable means:

Convert the expression to type void
Discard its value
Tell the compiler: this is intentionally unused
So for an unused parameter, this pattern avoids warnings like “unused parameter”.

Why it helps in your case:

variable = NULL writes to the parameter, so compiler can warn “set but not used”
(void)variable does not write anything, it only marks the parameter as intentionally consumed
Typical use:

(void)kwoObject;
(void)indexedKeyBlock;

It has no runtime effect in practice, it is only for clarity and warning suppression.

+++ src/Learning/MHHistograms/MHTruncationFloatingPointFrequencyTableBuilder.cpp
@@ -312,13 +312,13 @@ void MHTruncationFloatingPointFrequencyTableBuilder::ExtractFloatingPointBinBoun
 	assert(cIntervalLength < dTruncationBinaryEpsilon or
 	       (longint)(cIntervalLength / dTruncationBinaryEpsilon) == cIntervalLength / dTruncationBinaryEpsilon or
 	       nHierarchyBitNumber < GetMainBinHierarchyRootLevel() or
-	       cIntervalLength / dTruncationBinaryEpsilon >= LONG_MAX or
+	       cIntervalLength / dTruncationBinaryEpsilon >= (double)LONG_MAX or
 	       cIntervalLength == KWContinuous::GetEpsilonValue());
 	if (cIntervalLength < dTruncationBinaryEpsilon)
 	{
 		assert((longint)(dTruncationBinaryEpsilon / cIntervalLength) ==
 			   dTruncationBinaryEpsilon / cIntervalLength or
-		       dTruncationBinaryEpsilon / cIntervalLength >= LLONG_MAX or
+		       dTruncationBinaryEpsilon / cIntervalLength >= (double)LLONG_MAX or
 		       cIntervalLength <= KWContinuous::GetEpsilonValue());
 
 		// On ne traite que le cas d'un intervalle plus grand que le epsilon de la singularite en 0
